requirements.txt
iscc-sdk==0.8.0
pywikibot==10.0.0
python-dotenv==1.1.0
requests==2.32.4
PyJWT[crypto]==2.10.1
py-multihash==2.0.1
base58==2.1.1
SQLAlchemy==2.0.41
PyMySQL==1.1.1
alembic==1.16.1
pillow==11.3.0
